fix for: Encoding::CompatibilityError incompatible character encodings: UTF-8 and ASCII-8BIT
force encode column data with the encoding set in database.yml

When submitting a form, data fetched in a view from SQLAnywhere is encoded with ASCII-8BIT although database.yml is properly set up
The same code works with GET requests. Strangely not with PUT/POST.

I have to investigate, or need help with this.

Signed-off-by: Claudius Steinhauser <APL\claudius@cs-linux.apl.int>